requirements.txt
pytorch-lightning
adabelief-pytorch==0.2.0
git+https://github.com/PytorchLightning/pytorch-lightning-bolts.git@master